Add functions, use existing array concat
Add apply3 function to wrap an arity 3 function in an arity 1 function
that takes a 3 element array and calls the arity 3 function with each
element as an argument.

Revert back to the built-in array concat function and update the tests
to make sure the original arrays aren't modified.

Add a simple copy function that wraps array.slice(0).

Fix seqBy to go use descending or ascending arguments and positive or
negative step values.